requirements.txt
ipython==8.14.0
numpy==1.21.5
Pillow==10.3.0
pygame==2.5.2
matplotlib==3.8.0
do-mpc==4.6.4
ipykernel==6.29.4
